docs-translator-a2a/requirements.txt
# A2A VaaS Service Dependencies
# Core frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# CrewAI for agent orchestration (use installed versions)
crewai>=0.157.0
crewai-tools>=0.75.0

# LLM providers
openai==1.3.7
google-cloud-aiplatform==1.38.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
